Set the neutron default workers
As stated in neutron's default conf file:
  # This feature is experimental until issues are addressed and testing
  # has been
  # enabled for various plugins for compatibility.

This change has been shown to be reliable in manual testing of the gate
jobs. These jobs had been seeing transient failures possibly leading
back to this value. The changes here have been set within the
`bootstrap-aio.sh` script such that gating is using a consistent
environment even if these options change their values in the
future.

Thanks to Evan Callicoat for doing the work to isolate the failures.

Change-Id: I98116eef94a7240addfdf449d116ec1c24260c59
Co-Authored-By: Evan Callicoat <evan.callicoat@rackspace.com>
Closes-Bug: #1425255