Add neutron neutron_num_sync_threads variable
In the heat multi node test we were setting neutron_api_workers to 0.
We then noticed that the dhcp agent wasn't running correctly and Apsu
traced this back to num_sync_threads being set to 0 (currently, the
value of neutron_num_sync_threads will get set to neutron_api_workers if
neutron_api_workers is overriden).  While neutron's api_workers config
option can be set to 0, this is not a usable value for num_sync_threads.

This commit also changes a few instances of ``VCPUS`` to ``VCPUs`` in
documentation strings.

Change-Id: I0f8af934942e6ad58e0c9fd7bfad619d9e98a5bb
Closes-Bug: #1436916