Set neutron_lib_dir depending on neutron_venv_enabled
When neutron_venv_enabled is set to false, neutron fails to deploy as
it cannot find the alembic_migrations dir.  This is due to neutron
being installed in dist-packages outside of the venv and site-packages
inside the venv.  This commit adjusts neutron_lib_dir depending on
whether or not neutron_venv_enabled has been set.

Change-Id: I9d992993b116e85216c07d6fdb6a4f99398a2102
Closes-Bug: #1508041