Handle migrations for non-ml2 plugins
If a deployment uses networking-plumgrid then neutron-db-manage will
fail due to the plugin not having the necessary expand/contract
alembic migration branches.  This issue does not appear to be specific
to networking-plumgrid so for now we will assume that if
neutron_plugin_type is not 'ml2' then we have to run all migrations
with neutron-server down.

This patch can be reverted once all networking plugins use
expand/contract migration branches.

Change-Id: I6e1f921e64b8e8994dea4f36011d073a8eee92ef