Enable functional convergence testing
This commit adds test playbooks to deploy rabbitmq, galera, keystone,
and neutron for a functional convergence test. Note that we point at a
non-existent memcached server, but this can be adjusted to build
memcached if deemed necessary.

The openstack_services.yml file from openstack-ansible is downloaded and
included as extra vars to maintain consistent testing of OpenStack
project repo SHAs between this independent role and OpenStack-Ansible's
integration.

neutron_requirements_git_install_branch has been bumped to resolve the
issue of the required eventlet version no longer being available on PyPi
https://review.openstack.org/#/c/277912/

The test also checks if it's running in OpenStack-CI, and if it is -
then make the containers use the host's pip.conf and some of the apt
configuration.

Change-Id: I8905586727c327ac356b5a34897dbb096fd03e2f
Partial-Bug: #1553970