Remove iptables checksum rule script
https://review.openstack.org/#/c/148718/ has been merged so including
the post-up-checksum-rules script is no longer necessary. A new task has
been added so that the script will be removed during upgrades from Liberty.

Tests have been added to ensure that the dhcp agent is active, that the
dhcp network namespace is being created as expected, and that this
iptables rule is being created within that namespace.

The unconfined apparmor profile has also been applied to the neutron
test container so that it has the permissions required to create network
namespaces.

Change-Id: I068d091873d2744b0849b0d52a8083e129841b1b