Use internal endpoint for Nova notifications
Since https://review.openstack.org/#/c/254900/ has merged, Neutron
can now send Nova notifications over the internal endpoint. Since
the default is to use the public endpoint, it is not unlikely to
experience breakage with Neutron-Nova notifications when using SSL
termination on public endpoints. It is preferable to use the internal
endpoint for internal service communication.

Change-Id: I8bb55162cb22c0d8acb97903d7bd5766a9110982