removed duplicate key
In review of this role it was discovered that we had a duplicate
key within a task. While this is acceptable in Ansible as it will
just use the last key found in a given task it is wrong and should
be removed. In later version of Ansible (2.1.0) a warning is printed
stating something similar to the following:

   [WARNING]: While constructing a mapping from xxx.yml, line xx,
     column xx, found a duplicate dict key (changed_when).  Using
     last defined value only. [WARNING]: While constructing a mapping
     from xxx.yml, line xx, column xx, found a duplicate dict key
     (changed_when).  Using last defined value only.

This chage cleans up the role by removing the duplication.

Change-Id: I61ca269403f2293306980811a9f549ec4bd209a0
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>