Add options to auto enable VPNaaS
This change makes it possible for the neutron role to enable VPNaaS using
the reference OpenSwan driver by simply setting the `vpnaas` option within
the **neutron_plugin_base** list. This capability has been within the role
for some time however never specifically configured.

The change will bind the vpnaas agent to any location that is within the
neutron_l3_agent group. The agnet brings with it a new ini file and a new
entry point for config_template overrides. The vpnaas packages will be
installed when the agent is enabled within the target location of the
running agent services.

Change-Id: Ie9f89f7d2a942df40841d72c82f3a34315c6bed9
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>