Dynamically determine whether Neutron Agents should be enabled
Both the Nuage and PLUMgrid configuration guides advocate overwriting
the entire neutron_services dict in order to disable the DHCP,
Metadata and LinuxBridge Agents. This method is prone to error and
should ideally be more dynamic.

This patch implements a more dynamic approach based on the
neutron_plugin_type and the neutron_ml2_mechanism_drivers that are
set.

This will hopefully simplify the configuration of these and other
ML2 mechanisms (eg: openvswitch) when they're implemented.

Change-Id: I38e1fad4dfae24b32f4370b3da16561c42cd670d