Missing init script of VPNaaS
The playbook runs fine but it's missing the init
script of the neutron-vpn-agent which is necessary
to handle the new VPN services created.

Also there
is a wrong service_name "neutron-vpnaas-agent" instead
of "neutron-vpn-agent" (which are the binary that are
in the python libs folder).

Correction of the
neutron_driver_vpnaas using the neutron_vpnaas package
instead of the neutron package (deprecated).

vpnaas_agent.ini - added external_network_bridge and
interface_driver that were missing

Change-Id: I953f873bcdc17f2b78bb5753753b2f52d29082e1
(cherry picked from commit 9a3022d959c4c0d4d17bb0ddf58cc1c6eda35805)