Add metadata checksum fix for AIO-type networks configs
When running in an AIO, we need to implement an iptables rule in any
neutron_agent containers to that ensure instances can communicate with
the neutron metadata service. This is necessary because in an AIO
environment there are no physical interfaces involved in instance ->
metadata requests, and this results in the checksums being incorrect.

This is a necessary patch in order to work towards getting rid of the
run-playbooks.sh script in the integrated repository. With this patch
in place we will be able to set the AIO to activate this code path by
setting 'neutron_metadata_checksum_fix: True' in the AIO's
user_variables.yml, forgoing the needs to implement this in a bash
script.

Backport note:
This is being backported due to the numerous IRC conversations which
we have related to run-playbooks being used multiple times.
Implementing this backport will reduce the number of times we have to
field questions about lost container connectivity.

Conflicts:
>------defaults/main.yml (due to multi-OS in master)

Change-Id: I008bfdb2960800845703e721b38640b7434d1404
(cherry picked from commit e5726802927b638da821717e8dae24a723aeee53)