Cleanup metadata-proxy when old versions are present
The metadata proxy service will now be cleaned up when this role is
executed if an old version of the metadata-proxy is still running.
Once the old versions of the metadata-proxy have been cleaned up
the metadata agent will respawn the correct process within 60 seconds.

This change is required to ensure upgrades are always executing the
correct code from a given release.

Change-Id: I3a0b5c5b75742f06a94e5334c749b88e54f7f43c
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>