Implement standardised Ubuntu Cloud Archive repo vars
This patch implemented the vars for UCA which are present in the
os_nova and ceph_client roles in order to improve consistency and
easier overriding of any of the chosen parts.

The tasks implementing the UCA configuration are also slightly
adjusted to be a little more robust and complete.

Change-Id: I5ab9238245932de7f3ffc07ffe52725616f4f637