Remove static log-file parameter from init
The init/systemd scripts support dynamic parameters to be passed
to init services, however there is a static --log-file option that
was configured to be passed to ALL services using the init templates.

For services that are implemented which do not support --log-file
parameters this causes a startup failure ie.
calico-felix: error: no such option: --log-file

This change moves the --log-file parameter into the neutron_services
dictionary, allowing the parameter to be passed on a per-service basis.

Change-Id: I9e98f730e3d167feb40c2f7e898c10f53041bffd