Enable higher performance I/O through
In Ubuntu 16.04 the updated kernel that is shipped has conntrack support.
The inclusion of conntrack led to a new firewall driver that removed the
need for hybrid vif plugging. Hybrid vif plugging allowed security groups
to operate properly, but led to significant performance degradation's
when paired with OVS.

Since the inclusion of OVS for the PowerVM driver occurred during the
OpenStack Newton release, it aligned well with the Ubuntu 16.04 release.
As such, the PowerVM driver team did not enable hybrid vif plugging, but
instead focused on standard plugging.

This change set allows an option to turn off the hybrid vif plugging
(for all architectures), but will automatically detect it for PowerVM
and turn it off.  This will enable deployments on newer operating systems
to have higher network I/O performance.

Change-Id: I0801b0d6925a0d3c0ae9b14d38310d5906355b98
Partially-Implements: blueprint powervm-virt-driver