Check host group_names for group membership
If a task only applies when a host is a member of a particular group,
check for the group in a host's 'group_names' rather than checking for
the name of the host in the group itself.

These are functionally the same check but, for the latter, a fatal error
can occur if any particular group does not exist at all within the
environment.

Change-Id: Ib64ae9454e228da2ea37bc6d2760857f399ccade