Move UCA repo URL var to role defaults
In order to expose the var in role documentation and to allow the use
of dynamically set facts to override the value in CI environments the
variables are moved from the role vars to the role defaults.

The test preparation implements an override of the URL for OpenStack-CI
to make use of the local mirrors.

The variable to enable/disable the use of UCA has been standardised
to the same as the os_nova role.

Change-Id: I442f6dca8345f78848536f2fc923b72be8a6d092