Fix race condition on starting services
We're seeing a lot of failures for loading the neutron-linuxbridge-agent
on localhost in gating. There is a race condition that means the service
start is attempted before the configuration file is placed.

Although, this doesn't happen all the time - it suggests the static
inclusion on the init tasks may well be causing this error.

Change-Id: Ia9fbf588f64afe0e979163a319d3c07162ae8e17
Closes-Bug: #1628851
(cherry picked from commit 52965e146018b72a69b5e0350b087be04ea81ee9)