Move optional plugin installs out of the venv
The optionally enabled plugins do not need to be included in the
neutron venv, but should be added to the venv when enabled.

This prevents clashes between plugins and unnecessary db
migrations, including migrations which may not be working due to
changes in neutron core which haven't been applied to the plugin
yet.

Combined backport of:
- https://review.openstack.org/395556
- https://review.openstack.org/395570

Related-Bug: #1640319
Change-Id: I7dcab3f3d56118b93b9a5d74065234acb620ae22
(cherry picked from commit d4dd9df137877756507a5fcfa033d94f0f743429)