Move optional plugin installs out of the venv
The optionally enabled plugins do not need to be included in the
neutron venv, but should be added to the venv when enabled.

This prevents clashes between plugins and unnecessary db
migrations, including migrations which may not be working due to
changes in neutron core which haven't been applied to the plugin
yet.

Additionally, to unblock the CentOS-7 check, this patch includes
adding a task to remove any problem packages on the target host.

Related-Bug: #1640319
Co-Authored-By: Donovan Francesco <donovan.francesco@gmail.com>
Change-Id: I7dcab3f3d56118b93b9a5d74065234acb620ae22