Remove mistaken 'neutron_pip_package_state' variable usage
The backport https://review.openstack.org/395718 mistakenly
included the usage of the variable 'neutron_pip_package_state'
which was introdiuced in master for the purpose of the
updated ansible-lint checks. This variable is not valid in
stable/newton.

This patch removes the variable and sets the state to 'latest'
as with all other pip install tasks.

Change-Id: I09518913d221abc73904949dc6c780310edabc5f