Neutron upgrade testing.
Implement upgrade testing for os_neutron role.

This patch adds an upgrade testenv, which configures a previous version
of Neutron, and the test-vars from the previous stable branch to be used
by the previous version of Neutron.

We don't run functional tests after the initial stable/newton deploy
because a working stable/newton branch is already tested, and it's
assumed it worked if the upgrade works too - which is more
efficient.

This follows the pattern setup for Keystone in:
https://review.openstack.org/#/c/384269/

Change-Id: Id4b2cf8197ca902ab5b3d925175c3d7c698ec802
Implements: blueprint upgrade-testing