Fix calico test
Pins calico-felix agent at 1.4.5 because the 2.x tags and master
branch are a currently WIP rewrite of the agent in Go. We will need
to rework the install process for Felix before moving beyond the
1.4 stable tags.

Also set the tempest network settings to accommodate calico's local
network deployment (tempest role defaults are vxlan, and calico
eschews overlay networks.)

Set networking-calico branch to master since it is not versioned
with Openstack release cycles.

Change-Id: Idf4c558d9e2a31f2a22357ac12faaaa291de3a06