Use new ansible package module to install distro packages.
From Newton onwards we're able to take advantage of
the Ansible package module, instead of conditionally
executing the apt or yum module.

Implementing this is an optimisation which we can do
in master and backport to stable/newton,
reducing the execution time.

Reference: http://docs.ansible.com/ansible/package_module.html

Change-Id: I5727e2cb87e650d6348e46d120310b825af6127a
Related-Bug: #1642654