Clean up DVR conditionals
I543d1b5d3d50b6e7936605f436805d2f0ad9b3e9 removes the need for the
nova_compute based conditionals determining whether to enable L3
agent or metadata agent on the target. With the referenced change,
the hosts are only added to the group if they need the agents, so
we do not need these conditionals in the role.

Change-Id: I5dbafd3b68ddd45a4c44297c51412e092a5024cf