Allow override of the repo filename
Apt cannot have 2 mirrors with the same content in 2 different files.
If a deployer has an apt mirror with uca, the deployer
still need to add a repository, but will also need to define the
filename used, in order to avoid clashes.

This commit makes possible to decide the filename for the repo.

Change-Id: I05c57fe1e43c15757797ff4ad6a2acf39441cb50
Signed-off-by: Jean-Philippe Evrard <jean-philippe.evrard@rackspace.co.uk>