Keep the N behavior optionally
This commit makes possible to allow the definition of a filename for
the repositories, but also keeps the current Newton behavior, in order
to be backported. This would make updates/upgrade more flexible, and
is in line with what has been done in other repos, like galera_server.

The previous commit on these forced a "non-change", and made the
variable mandatory. This reverses it, while it keeps the functionality.

Change-Id: I1d126de5b911fa6600f30d5c9d6a49d6f78564ec
Signed-off-by: Jean-Philippe Evrard <jean-philippe.evrard@rackspace.co.uk>