Allow override of the repo filename
Apt cannot have 2 mirrors with the same content in 2 different files.
If a deployer has an apt mirror with uca, the deployer
still need to add a repository, but will also need to define the
filename used, in order to avoid clashes.

This commit makes possible to allow the definition of a filename for
the repositories, but also keeps the current Newton behavior.
This would make updates/upgrade more flexible, and
is in line with what has been done in other repos, like galera_server.

Manual combined backport of:
- https://review.openstack.org/#/c/417946
- https://review.openstack.org/#/c/416245

Change-Id: I05c57fe1e43c15757797ff4ad6a2acf39441cb50
Signed-off-by: Jean-Philippe Evrard <jean-philippe.evrard@rackspace.co.uk>