Ensure neutron_bin gets created
The `neutron_get_venv` check is causing the creation of the
`neutron_bin` directory to be skipped on all hosts. This causes the
unarchive venv step to fail immediately afterwards.

This patch corrects the check in the `when:` argument and ensures that
the directory is created to hold the venv content.

Closes-Bug: 1657138
Change-Id: Ia7a91c8b7c93affe8e378ce933299b050c63e854