Skip with_first_found when no files are found
The upgrade tests we run in Ocata will run against stable/newton where
the Redhat7 vars file does not exist, since we don't have Redhat7
support.

Due to the way includes happen, the files are still evaluated, which
means the task will fail. To avoid this we allow the task to be skipped
if no file is found.

After stable/newton the Redhat7 vars will be found, so we don't need
this change in Ocata+.

Change-Id: Idb74e6e05c20a85354d5b1acd12cdc90d7cda2f5