Work around Trusty CI bindep issue
Ubuntu Trusty is not a supported platform for Ocata onwards.
Currently the CI system is executing jobs against Trusty
unintentionally and it's taking some time to unwind it.

This bypasses the attempt to install a package that isn't
available for Trusty.

Change-Id: I343e357fa036222b056939d8db9cd8f7e8a45b7a
(cherry picked from commit 228158e91be8338b98c7cdf630eb40d2707d8141)