Update test set and use test plugins
Previously the test set used was the default set in
the os_tempest role, which are scenario tests. Those
tests cannot run without more services deployed.

This patch implements the use of API tests and the
tempest plugins from the Neutron repository.

Additionally, the postgresql library is removed as
it is no longer required.

Change-Id: I4710511d7824ad031cafb4dac5521c0b1b34931e