Cleanup plugin loaded_base
neutron_plugin_loaded_base is only used in the template. This
var probably shouldn't be used by any deployer, so we mask this
variable from the defaults. Currently, we still keep this
overridable behavior: neutron_plugin_loaded_base can still
be overriden by a deployer, and we still rely on plugin_base if
the deployer hasn't defined it.

Change-Id: I8d7903dbb9a6f7d096eb3f840164b87c48eb5e4d