Only use handler on filtered services
Since we create a filtered list of services we should use the handler to
only restart those services, rather than all services.

Additionally, we no longer need to check "service_en" since that will
happen when the services are filtered. We can then check if the service
exists before performing the metadata-proxy cleanup.

Change-Id: I4ca3d90123e146d09cab20a0eda627014d9717e6
Depends-On: Ic152cfb01930d8bde10ea6d9aa5bba173a87e376