Move static defaults to vars
Defaults used to contain a lot of variables for internal
purposes, not to be defined by the deployers or our CI
systems.

This doesn't respect the idea of defaults/main.yml and
vars/main.yml explained in the following link [1].

Because this role's variables were in a non-organized
order, a general cleanup was necessary.

The following structure was taken:
- Verbosity Options (only in defaults)
- Packages Options
- Python code details
- Generic neutron config
- Tunable Overrides (only in defaults, sorted alphabetically)
- Quotas
- DB (Galera) integration
- RPC (Rabbitmq) integration
- Identity (Keystone) integration
- Telemetry integration
- Other external roles integrations like Designate
- Plugins Loading
- Plugin x Configuration (ML2, L3, DHCP, Metadata, ...)
- Internals: neutron_services mappings
- Internals: driver mappings
- Internals: convenience vars for py_pkgs

[1]: https://docs.openstack.org/developer/openstack-ansible/
     developer-docs/contribute.html#ansible-style-guide

Change-Id: I431ec3dd834a5624fb1479cc18ece045198431a1