Ensure that services restart in a particular order
Currently when multiple services share a host, the
restart order is random. This is due to an unordered
dict being used to facilitate the mapping of services
to their groups, names and other options.

Based on [1], this patch implements changes to the role
to ensure that services on the same host are restarted
in the correct order when the software/config changes.

[1] https://docs.openstack.org/developer/neutron/devref/upgrade.html

Conflicts:
>------vars/main.yml

Change-Id: I368b51ef37763f4163ead591d6743c4d56962ef9
(cherry picked from commit 2f10f754252b9e0a14dc2f9431bea9994e959866)