Simplify database migrations
Instead of using a module to check for migrations,
we now execute the expand every time the venv is
updated. We also implement a simple check for
offline migrations and do the proper server shut
down and execution as per [1].

The integrated playbook will target all
neutron-server hosts at once when executing the
role to make sure that the db contract is done
with them all offline.

[1] https://docs.openstack.org/developer/neutron/devref/upgrade.html

Change-Id: I7dfef45e1ed8a8dfa464f1b3d20c90ae3348ce2e