Resolve migration issues
A run_once task decides on a host without the
context of conditionals related to the task.
The filter is done prior to evaluating any
conditionals.

Also, only the neutron-server neutron.conf has
the database connection credentials as a
security measure, so it is not able to execute
the database migrations.

This patch ensures that the database migrations
are only run once, but are also executed
wherever neutron-server is. It also adds
multiple instances of neutron-server and
neutron-agents to better test against problems
of this nature happening again.

Additionally, the patch uses the handlers for
the offline migrations to ensure that
neutron-server is only restarted once when a
new tag is deployed.

Conflicts:
>------tests/ansible-role-requirements.yml
>------tests/test-neutron-resources-upgrade.yml

Depends-On: I5cbb3824430dc09b36476f81e0cdfd4f0a15f497
Change-Id: I672ceb0848415c8f2653ebc8f7556db77f7f001c
(cherry picked from commit afe63f6f1972f6747134544d8b2edea608da4f37)