Cast the resulting intersect set into a list
When using Ansible with python3, the result of
the intersect filter is a set, not a list. This
causes a failure when trying to access item 0
in the list.

In this patch we cast the set to a list before
accessing item 0. This will work for both
python2 and python3.

Change-Id: I9b2a01882bfd5f02e0e0b85e1d57fab0fa88d0e9