Switch from run_once back to inventory scoping
As it turns out, run_once executes a task
once per batch when executed from a serialised
playbook instead of once per set of play hosts
as previously thought.

In this patch we implement a combination of
dynamic inclusion, inventory scoping and
play host scoping to achieve the required
goal of only running it once, even when the
playbook is executed using limits.

Backport note:
ansible_play_hosts in Ansible 2.1 provides
the hosts for the play serial batch and not
the hosts for the play as a whole. It is
therefore not suitable, so we leave the
task scoping as it was.

Depends-On: I548971e2de92280fe2cc4ff0a6a926497733fa7d
Change-Id: Icde03cd2610f52859fd8d2d7540f442a36e51695