Allow only vxlan networks to be created in Neutron.
If network_mappings is defined but empty, we should treat it as though
it is undefined. Otherwise the template tasks will fail due to there
being no "network_mappings[0]".

We should add a test case for this once the fix is backported to the
stable/ocata branch, to avoid a regression in future. Currently this
would fail the upgrade gate due to this bug coming up in stable/ocata
branch.

Change-Id: Ifd0d2c23bb65225d27c74360ff521b84676b7a81
Closes-Bug: #1528307