Fix VPNaaS variable definition for non-ubuntu
VPNaaS variables are defined for ubuntu, but not the other
distros.

One choice was to move these vars into vars/main.yml, but
would have prevented the easy overriding of it.
Because it's a driver, there are chances that someone wants
to override it, so it should be in defaults.

Closes-Bug: #1711577

Change-Id: I843e950e5c668da88500d60cf87bb9d19246262d