L3 Support for ODL deployments
The current ODL deployment does not correctly support L3 when ODL implements
the L3 logic (instead of using neutron-L3-agent)

Note that the neutron_plugin_base variable is a list with the neutron plugins
to be used and it is read by neutron.conf to list them. If this variable has
odl-router or odl-router_v2, neutron will delegate the routing logic to ODL
and will not start neutron-L3-agent.

Change-Id: I819e566c192ab1a00736f0aa9d7d04939f84f66d