Fix role tests
1. Update vars and test tooling for Pike

The etcd role pin is set to the current pin in the
integrated repo.

The networking-calico driver is set to use the latest
tagged release.

The dragonflow repo is left using master at this time
with the intent of switching over to use stable/pike
when it is created.

2. Update the tempest plugin version.

Neutron needs to have its tempest test from its
stable/pike branch, not from master branch.

This plugin needs a bumped sha (17.0.0), which comes
from the tests repository (see depends-on).

3. Bumped linux bridge filters.

Updated paste, policy and rootwrap configurations
from the 2017-09-28, this way we are in-line with
neutron stable/pike files.

Depends-On: I7732c4f7ac44ad7da95df73c9d1df9e08cbd8d21
Change-Id: I1489491a4acef9b32e2d62a46c9e17ea2507314a