Work around a neutron dns_domain transition issue
Neutron changed the way DNS domains are handled in newton from using
dhcp_domain to dns_domain.  This transition actually happened in the
stable/newton release but was disclosed in the stable/ocata release notes[1].

This issue was corrected in the stable/ocata branch of
openstack-ansible-os_neutron[2] but was not corrected in stable/newton.

This patch works around the odd transition timing in stable/newton by
also applying the new dns_domain setting in the neutron.conf with the value
from the legacy neutron_dhcp_domain variable that exists in stable/newton
OSA.

[1] https://docs.openstack.org/releasenotes/neutron/ocata.html#upgrade-notes
[2] https://github.com/openstack/openstack-ansible-os_neutron/commit/d9cc60b9ce200e24f69c31001bf40151f03e1e73

Change-Id: I11846880d88d8b0d0b7236d5825be31525c1d6b0