Add git and iptables to the list of packages
1) When doing an upper branch upgrade test (anything above Ocata), we
clone previous branch, and run a playbook that targets this previous
branch. It's not possible to merge anything with the upgrade job
failing.

2) The upgrade jobs are currently failing in neutron, because pip install
in the role, for the developer mode, requires git and pip installed
INSIDE the container. We therefore need to add git to the list of packages
to install [1]

3) Similarily, iptables need to be installed on all the agents, so
we add it to the list of the packages to install.

[1]: ```
Collecting neutron from git+https:
//git.openstack.org/openstack/neutron@stable/pike#egg=neutron
(from -c /opt/developer-pip-constraints.txt (line 1))
Error [Errno 2] No such file or directory while executing command
git clone -q https://git.openstack.org/openstack/neutron
  /tmp/pip-build-Lk_L35/neutron
Cannot find command 'git'
```

Change-Id: I1edfae74fc62d523d670bf5932587aa6626512d2
Co-Authored-By: Manuel Buil <mbuil@suse.com>