Update upgrade role for Queens from P
This changes the a-r-r with the proper version, and fixes
the repo path for role upgrades:

The role will, during test-upgrade-pre, have its current branch
checked out, and will deploy the current branch infrastructure,
including the generation of the constraints.

Then, the installation of previous branch of the role will use
the constraints for its previous version, but will have no
constraints for it, and fail.

We need to generate the constraints for the previous version too.
This should do it.

On top of it, we need to enfore git and iptables to the list of
packages, because of the issues happening in the developer mode:

Error [Errno 2] No such file or directory while executing command
git clone -q https://git.openstack.org/openstack/neutron
  /tmp/pip-build-iS3zW0/neutron
Cannot find command 'git'\n"

This patch also implements an initial set of jobs intended to
match the current job execution method. It does not intend to
improve how the jobs are executed - only to replicate what is
currently in openstack-infra/openstack-zuul-jobs and provide
the platform to iterate on.

To avoid the chicken and egg issue, this patchset temporarily disables
the upgrade testing.

Co-Authored-By: Manuel Buil <mbuil@suse.com>
Change-Id: I48a554e4ef92bc3c156e43881013cc85df0dc8d4