Used cached git sources and enable depends-on
Many, many patches have gone into the master branch to
enable the use of depends-on for role tests and to make
use of the git clones put into place by zuul. This patch
ensures that this functionality is now available for this
branch too.

The primary benefit for doing this would be to make use
of the cached git clones provided by zuul to improve test
reliability.

The secondary benefit would be to have the ability to
use depends-on to test patches together.

However, this will come at the cost of requiring a tox
configuration change, and the implementation of the
tests/tests-repo-clone.sh and run_tests.sh scripts
across all repositories.

The jobs are adjusted to match those present in the
last jobrun by jenkins as shown in:
https://review.openstack.org/511309

Change-Id: Id86b2848125de1261f7bcbe39bee4674f32c3861
Depends-On: I2abda86df19980c963e3cb176db2f877ba9a312b