Fix tempest tests
The neutron plugins are now in a new repository, and therefore
the path to the whitelist has changed.

On top of that, the "when:" condition on the include playbook
naturally didn't trigger. There is no when on play(book)s!
Due to the fact this was ignored, we were running the tempest role
twice for a while. It didn't break, because the facts were properly
gathered, but it was suboptimal. If facts failed to gather, the
os_tempest role that gets included in the dragonflow play would fail,
and that would break the functional test (not only the dragonflow test).

The "when:" condition on plays should be removed with ansible 2.4 coming
soon anyway.

Change-Id: I4e4110683182310ca662de49ca4437ca56f9ca4c