Fix LBaaSv2 config and package requirements
The LBaaSv2 (not Octavia) requires the net-utils(tools) package to
function. Additionally the user_group needs to be set to an OS specific
value. This change adds a default allowing the user to override the new
option `neutron_lbaasv2_user_group` as needed but the default will be
inherited from the os specific variable file.

Change-Id: Ic4b9461e369f1db94dfcfc3cebae034b3d38bb09
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>