tasks: Ensure Open vSwitch is started for all providers that need it
Previously, openvswitch was only started when we were using the ODL
provider but we should start it for all providers that need it. As such
we move the conditional which checks whether Open vSwitch is needed into
a new variable and we use that to start/enable the service and also to
install the package when needed.

Change-Id: I4ec4fc2c2124a941f24478cbdf39ed34c0dce56f
(cherry picked from commit 4b940c8337017dcfb275f3eba90f4e8f5fa6dffc)