Add SELinux contexts for neutron log directory
This is a backport of:
- https://review.openstack.org/543588
- https://review.openstack.org/545503

The log directory for neutron has the default_t SELinux context and this
prevents rsyslog from accessing neutron's logs. This patch ensures that
the file contexts are set properly for neutron's logs.

This change also makes neutron's log directory configurable using the
`neutron_log_dir` variable.

Closes-Bug: 1748968
Change-Id: Ifbcca131435c8963cc9c1b85c000cc040fab27ab