neutron_needs_openvswitch should be jinja, not a string
In I4ec4fc2c2124a941f24478cbdf39ed34c0dce56f the long conditional set
was broken out into its own variable to make it easier to grok, but
the new variable was not wrapped in jinja braces, so it's evaluated
as a string instead of jinja to interpret.

Also, all conditionals that use the variable should be checking for
a boolean value. Right now it might just consider the result true
because the variable exists, rather than the value of the variable.

Change-Id: Ic5210d62412ccf08e7d71860159aaaf37be9ddfd