Trim whitespace in Neutron configuration file for LBaaS
If neutron_lbaasv2 is set to true, the if statement is evaluated
and an extra 2 spaces are added.  This results in two spaces being
added before the '# General, applies to all host groups' which
results in config_template being unable to parse the file in
addition to Neutron failing to parse the configuration as well.

Change-Id: I4e04e1b9e6219d60c703240d01d6bf8a1a32ccda