Set new parameter in OVS to support ODL Oxygen
ODL oxygen provides a new framework to configure tunnels called autotunnels.
SFC requires tunnels to be flow-based and to arrange that in the autotunnel
framework, the OVS internal parameter of-tunnel must be set to true.

That parameter is not read in previous ODL versions. Therefore, this patch
does not break backwards compatibility.

Change-Id: Iec6a741d1e1081da547fe6fc0a67dd723e714d60
Signed-off-by: Manuel Buil <mbuil@suse.com>