Do not pip install keystoneclient on the host
The keystoneclient package is being installed on the host by PIP but
that means that a whole bunch of required dependencies are being pulled
in as well.

This brings the host to a rather messed up state when installing
keystone from distro packages, since distribution and
PIP packages are being mixed together. Moreover, we do not need to
fetch the httplib2 module from PIP since it's provided by distribution
repositories.

Implements: blueprint openstack-distribution-packages
Change-Id: Ic40f0ac06ba63bc601dc0bc1cf3fafb87c88269f