Add osprofiler to api-paste pipeline
The osprofiler filter is included in the standard pipelines for
nova, glance, cinder, heat, magnum [1], but is missing for
upstream neutron.

This means when a deployment uses profiling, there is a gap in the
data collected when the neutron API is executing.

Upstream fix https://review.openstack.org/#/c/583362/

[1] https://github.com/openstack/openstack-ansible-os_nova/blob/master/templates/api-paste.ini.j2#L31-L32

Change-Id: I63e2e8a1886527d363986bbb82c069eaaaa4406a
Partial-Bug: #1781619