Update permissions on execution path folders
The execution and configuration path are being craeted by the role and
if the role is the first to create the the permissions will be 750 and
could restrict other services, like nova, from starting. This change
updates the permissions to 755 which will ensure normal read and
execution permissions are set.

Change-Id: I2d6e533392447a8096b7c5cd5ca453c78c65d1ab
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>