Move source build vars back to vars/main
All other roles have their pip_packages in defaults/main.yml
and the lists were all in vars/main.yml before [1] was
implemented. Since then, the py_pkgs lookup does not recognise
that neutron has the project_group of 'neutron_all', but
instead thinks it is in 'all'. This mismatches the git
repo project group, resulting in a failed repo build if the
group is not populated. It also results in neutron packages
building even if neutron is not configured to be part of the
build.

This patch moves the vars back to vars/main.yml to restore
proper functioning of the repo build and correct association
with the project group.

These vars are not used by the distro install process, so it
does not matter where they're sourced.

[1] https://review.openstack.org/573661

Change-Id: If8aa2dba6fc5b09a71c93de7f048ea716fc6a2c1