Drop SELinux support for CentOS 7
We do not have a maintainer at the moment for SELinux and hopefully
we will adopt the upstream openstack-selinux package, but for now
in order to let deploys in environments where SELinux is set to
permissive work, we'll have to remove these bits.

This change can be reverted whenever we have a maintainer that's
available to do the work required.

Closes-Bug: 1792050
Change-Id: I4c7b6a9c0d8ec1458a9396422d047e1327bb4d45
(cherry picked from commit 084559b8cf7c7fe2b7e6a530e6f6da804d38e056)