Fix api-paste.ini path in RedHat ditro install
This patch aims to revert the commit meant to
fix api-paste.ini [1] that has no effect to
solve the path problem in distro installs based
on RedHat.

After the neutron_install.yml task responsible
to install neutron role packages, some config files
are placed in /usr/share/neutron, some of these files
are soft links from /etc/neutron from /usr/share/neutron.

One of the config files that is used by neutron-server is
the api-paste.ini that needs to be placed in /etc/neutron in
order to the service to initalize correctly.

This is a workaround only for RedHat distro installs
and the root cause is still being investigated.

[1] https://github.com/openstack/openstack-ansible-os_neutron/commit/c0b17e9d89d86d0f9543c401316727cfcd6045f2

Change-Id: Ie5441b8d2b147bf0e57928a2b6ea6a745a96c67f