Provide support for "force_metadata" in dnsmasq
Metadata server is important in some deployments and VMs need a way to
reach it. Normally, neutron routers include a static route to the
metadata server and things work, however, when using external L3 plugins
(e.g. ODL) that static route might not be there by default and thus VMs
don't have a way to contact the metadata server. For these cases, there
is an alternative in which the DHCP server pushes the route to the
instances through the 121 message defined in the DHCP standard. To use
this alternative in dnsmasq, we need force_metadata = True in its config

Change-Id: I2def87c8e7ed4b9ac87b66f825f2766a67a9a802
Signed-off-by: Manuel Buil <mbuil@suse.com>
(cherry picked from commit 012329cb352a18b61abb1b028911ab621eb844f7)