Fix whitespace in neutron.conf template
The Jinja2 templating code can introduce whitespace which ends up in
front of the `# General, applies...` comment. This leads to a parsing
error when Neutron is started and the Neutron installation playbook
fails. Note that this change leaves _more_ whitespace in the beginning
of neutron.conf, but preserves the newline in front of the comment.

Change-Id: Ib9a9e44f25dcabb92df6067852e91115d30fe740