tasks: neutron_install: Fix group for neutron hosts
Previously, we only executed the AppArmor configuration when the
inventory_hostname matched one of the hosts in the neutron_appamor_hosts
list. However, there are cases were a multiple LXC hosts share the same
physical host. In that case, since we only record one container per
physical host, that container may not be part of the running inventory,
so the AppArmor configuration is never executed. We fix that by simply
iterating the neutron_all group which should contain all the hosts which
participate in the neutron deployment.

Change-Id: Ib523de857a5de9adc86bd90fe6919af17704192f