Adds ability to set no-resolv for dnsmasq.
Introduces variable neutron_dnsmasq_noresolv. When it is set to true,
no-resolv option will be placed in dnsmasq-neutron.conf.

If dnsmasq_dns_servers are not set, and neutron_dnsmasq_noresolv is True
dnsmasq will reply with empty response on DNS requests,
which may be used as a workaround of
https://bugs.launchpad.net/neutron/+bug/1501206
As a con, resolver is actually won't work at all, so all networks w/o
configured DNS will stop resolving DNS.

Change-Id: I7764e0d6100a8327f84f78da64f10f8e1fa355b1