Set appropriate default for vxlan multicast ttl
Currently no default is set which means vlxan multicast traffic
is dropped at the first L3 router. This patch increases the default
to 32 which is sufficient for any reasonable network fabric and
can be overridden if necessary.

Without this patch vxlan project networks on a routed underlay network
are not functional

Change-Id: Iddf2e412b3d1f23362d94a7eabd8c062d68ae287
Closes-Bug: 1755790