Convert dynamic includes to static imports
When task files are included using include_tasks, tags are not passed to the
included tasks. As a result, tags like neutron-config do not have the intended
effect. This patch changes include_tasks to import_tasks for all cases where
dynamic vars or loops are not used so that tags are properly handled.

Reference -
https://docs.ansible.com/ansible/latest/user_guide/playbooks_reuse.html
https://bugs.launchpad.net/openstack-ansible/+bug/1815043

Change-Id: I525265255b19018536365272f5a48baa486f8a0c