Create symlink to neutron-keepalived-state-change
Neutron tries to hit up neutron-keepalived-state-change bin
and looks for it in /usr/local/bin.

The workaround for this is to create a symlink of the bin
file inside the venv to the place neutron expects.

Removed condition on neutron_install_method for Queens as installs are
always from source.

Change-Id: Id3ca9b81d946a07d74157578981c78b2e35f02ff
(cherry picked from commit 1ea1e654a3321882e6cc9ef836085514eefc13bc)