Stop installing python-httplib2
In https://review.opendev.org/#/c/575701/, we installed python-httplib2
because we needed it for python-keystoneclient to work. However,
we no longer install python-keystoneclient as of
https://review.opendev.org/#/c/627791/. Except, we never stopped
installing python-httplib2 even though it is no longer needed.

This currently breaks CentOS because python-httplib2 is not present
in the base or updates repos that we enable, only in extras.

Change-Id: Iae15e90e0e572db806c712a60b8ea3794b9e9a71