Make venv download configurable
The neutron venv download configuration is set in vars/ instead of
defaults/ which is completely inconsistent from the nova role. This
implementation is suboptimal because it disallows configuration of
the venv download flag in group vars.

This update will make the neutron venv download vars consistent
with the nova configuration.

Change-Id: I1c2516fb2c95a8253271eaf626ccc923b2ee0846