l3 agent on network_hosts do dvr_snat, anywhere else dvr
With the change in queens of moving the neutron agents from
containers onto the network_hosts, if those hosts happen to
also be your nova_computes, VM's without a floating ip would
lose north/south network.

Checking for physical_host_group could make this patch work for
neutron_agents_containers on pike aswell, if we want to
backport it that far.

Change-Id: I6bb24a8719b3902d6f510c70705cdb63076a1a08
Signed-off-by: Erik Berg <openstack@slipsprogrammor.no>